+++ animal_shogi/board.py
@@ -1,42 +1,35 @@
-from copy import copy
 import json
 import re
+from copy import copy
 
-
-def simplify(board):
-    b = str(board)
-    b = b.replace(' ', '')
-    b = re.sub(r'^$\n', '', b, flags=re.MULTILINE)
-    b = b.strip()
-    return b
-
-
-def equals(board1, board2):
-    board1 = str(board1)
-    board2 = str(board2)
-    return simplify(board1) == simplify(board2)
+from .exceptions import ParsingError
+from .koma import LION, GIRAFFE, ELEPHANT, CHICK
+from .utils import equals, normalize_board_dict
 
 
 class Board(object):
     """
     +---+---+---+
-    |1,1|1,2|1,3|
+    |1,1|2,1|3,1|
     +---+---+---+
-    |2,1|2,2|2,3|
+    |1,2|2,2|3,2|
     +---+---+---+
-    |3,1|3,2|3,3|
+    |1,3|2,3|3,3|
     +---+---+---+
-    |4,1|4,2|4,3|
+    |1,4|2,4|3,4|
     +---+---+---+
     """
     SIZE_ROW = 4
     SIZE_COL = 3
 
+    # e.g. ' * +HI -OU'
+    BOARD_ROW_STR_PATTERN = re.compile(r'(\*|\+[A-Z]{2}|-[A-Z]{2})SIZE_COL'.replace('SIZE_COL', str(SIZE_COL)))
+
     def __init__(self):
-        self.my_mochigoma = {}    # type: Dict[Koma, int]
+        self.my_mochigoma = {}  # type: Dict[Koma, int]
         self.your_mochigoma = {}  # type: Dict[Koma, int]
-        self.my_board = {}        # type: Dict[Tuple[int, int], Koma]
-        self.your_board = {}      # type: Dict[Tuple[int, int], Koma]
+        self.my_board = {}  # type: Dict[Tuple[int, int], Koma]
+        self.your_board = {}  # type: Dict[Tuple[int, int], Koma]
         self.my_turn = True
 
     def show(self):
@@ -49,14 +42,18 @@ class Board(object):
 
         {
             'your_mochigoma': [],
-            'my_mochigoma': [{'koma': 'FU', 'num': 1}],
-            'board': [{'koma': 'OU', 'y': 2, 'mine': True, 'x': 1},
-                      {'koma': 'HI', 'y': 3, 'mine': True, 'x': 1},
-                      {'koma': 'KA', 'y': 1, 'mine': True, 'x': 1},
-                      {'koma': 'FU', 'y': 2, 'mine': True, 'x': 2},
-                      {'koma': 'OU', 'y': 2, 'mine': False, 'x': 4},
-                      {'koma': 'HI', 'y': 1, 'mine': False, 'x': 4},
-                      {'koma': 'KA', 'y': 3, 'mine': False, 'x': 4}]
+            'my_mochigoma': [
+                {'koma': 'FU', 'num': 1}
+            ],
+            'board': [
+                {'koma': 'OU', 'y': 2, 'mine': True, 'x': 1},
+                {'koma': 'HI', 'y': 3, 'mine': True, 'x': 1},
+                {'koma': 'KA', 'y': 1, 'mine': True, 'x': 1},
+                {'koma': 'FU', 'y': 2, 'mine': True, 'x': 2},
+                {'koma': 'OU', 'y': 2, 'mine': False, 'x': 4},
+                {'koma': 'HI', 'y': 1, 'mine': False, 'x': 4},
+                {'koma': 'KA', 'y': 3, 'mine': False, 'x': 4}
+            ]
         }
         """
         board = []
@@ -65,11 +62,15 @@ class Board(object):
                 x, y = xy
                 board.append(dict(x=x, y=y, koma=koma.yomi, mine=mine))
 
-        return {
-            'my_mochigoma': [{"koma": k.yomi, "num": v} for k, v in self.my_mochigoma.items()],
-            'your_mochigoma': [{"koma": k.yomi, "num": v} for k, v in self.your_mochigoma.items()],
+        my_mochigoma = [{"koma": k.yomi, "num": v} for k, v in self.my_mochigoma.items()]
+        your_mochigoma = [{"koma": k.yomi, "num": v} for k, v in self.your_mochigoma.items()]
+
+        d = {
+            'my_mochigoma': my_mochigoma,
+            'your_mochigoma': your_mochigoma,
             'board': board
         }
+        return normalize_board_dict(d)
 
     def to_json(self, path_or_fp=None):
         # Return dict
@@ -86,27 +87,37 @@ class Board(object):
     def __eq__(self, other):
         return equals(self, other)
 
-
     @staticmethod
     def _parse_row(row):
         r_value = []
         for i in range(Board.SIZE_COL):
             if row.startswith('*'):
                 r_value.append(None)
-            
 
     @classmethod
     def from_str(cls, s):
         b = Board()
-        current_x = 0
+        current_y = 0
         for row in s:
             s = s.replace(' ', '')
             if 'my_turn' in s:
+                # turn
                 b.my_turn = True if 'true' in s.lower() else False
             elif s.startswith('*') or s.startswith('-') or s.startswith('+'):
-                current_row += 1
-
-
+                # board
+                current_y += 1
+                m = cls.BOARD_ROW_STR_PATTERN.match(s)
+                if m is None:
+                    raise ParsingError('Illegal board pattern: {}'.format(s))
+                for x in range(cls.SIZE_COL):
+                    koma = m.group(x + 1)
+                    if koma != '*':
+                        mine = koma[0] == '+'
+                        koma = koma[1:]  # '+HI' -> 'HI'
+            # 'board': [
+            #     {'koma': 'OU', 'y': 2, 'mine': True, 'x': 1},
+            # elif s:
+            # # mochigoma
 
     def __str__(self):
         s = []
@@ -159,18 +170,19 @@ class Board(object):
 
     @property
     def possible_nexts(self):
-        """Returns Iterable[Board]"""
-        # make sure current_oard.my_turn is True
+        # type: () -> Iterable[Board]
+        """Returns possible legal moves"""
+        # make sure current_board.my_turn is True
         current_board = self if self.my_turn else self.flip()
 
         # 1. use mochigoma
         for koma in current_board.my_mochigoma:
-            for r in range(1, self.SIZE_ROW+1):
-                for c in range(1, self.SIZE_COL+1):
-                    if current_board._blank(r, c):
+            for y in range(1, self.SIZE_ROW + 1):
+                for x in range(1, self.SIZE_COL + 1):
+                    if current_board._blank(x, y):
                         board = current_board.copy()
                         board.my_turn = not board.my_turn
-                        board.my_board[(r, c)] = koma
+                        board.my_board[(x, y)] = koma
                         if board.my_mochigoma[koma] == 1:
                             board.my_mochigoma.pop(koma)
                         else:
@@ -213,100 +225,6 @@ class Board(object):
         return (1 <= r <= self.SIZE_ROW) and (1 <= c <= self.SIZE_COL)
 
 
-class Node(object):
-    def __init__(self, board, depth=3):
-        self.board = board
-        self.depth = depth
-        self._score = None
-        self.best_moves = None
-        # self.last_move  # TODO: implement
-
-    @property
-    def score(self):
-        if self._score is None:
-            raise RuntimeError("Run evaluate() first")
-        return self._score
-
-    @property
-    def child_nodes(self):
-        if self.depth == 0:
-            return []
-        return [Node(board, depth=self.depth - 1) for board in self.board.possible_nexts]
-
-    def evaluate(self):
-        """
-        - Calculate score
-        - Store the best moves
-        """
-        # Return the score of the board if this doesn't need to check further
-        if self.depth == 0:
-            self._score = self.board.score
-            self.best_moves = [self.board]
-            return
-
-        # Else check child nodes and take the max if it's my turn, else the min
-        if self.board.my_turn:
-            minmax = max
-            best_score = -1 * 10 ** 200
-            best_node = None
-        else:
-            minmax = min
-            best_score = 10**200
-            best_node = None
-        for n in self.child_nodes:
-            n.evaluate()
-            s = minmax(best_score, n.score)
-            if s != best_score:
-                # updating..
-                best_node = n
-                best_score = s
-        self._score = best_score
-        self.best_moves = [self.board] + best_node.best_moves
-
-    @property
-    def possible_nexts(self):
-        """
-
-        :return: iterator of Node
-        """
-        for board in self.board.possible_nexts:
-            yield Node(board, self.depth-1)
-
-
-class Koma(object):
-    def __init__(self, directions, score, yomi):
-        self.directions = directions
-        self.score = score
-        self.yomi = yomi
-
-    def __repr__(self):
-        return '<Koma {}>'.format(self.yomi)
-
-
-LION = Koma([[-1, 1], [-1, 0], [-1, -1], [0, 1], [0, -1], [1, 1], [1, 0], [1, -1]], 100, 'OU')
-GIRAFFE = Koma([[-1, 0], [0, 1], [0, -1], [1, 0]], 4, 'HI')
-ELEPHANT = Koma([[-1, 1], [-1, -1], [1, 1], [1, -1]], 4, 'KA')
-CHICKEN = Koma([[0, -1], [1, -1], [1, 0], [0, 1], [-1, 0], [-1, -1]], 4, 'TO')
-CHICK = Koma([[0, -1]], 3, 'FU')
-
 INITIAL_BOARD = Board()
 INITIAL_BOARD.my_board = {(1, 4): ELEPHANT, (2, 4): LION, (3, 4): GIRAFFE, (2, 3): CHICK}
 INITIAL_BOARD.your_board = {(3, 1): ELEPHANT, (2, 1): LION, (1, 1): GIRAFFE, (2, 2): CHICK}
-
-
-def main():
-    b = INITIAL_BOARD
-    while True:
-        b.show()
-        n = Node(b, depth=6)
-        n.evaluate()
-        print 'Score: {}'.format(n.score)
-        print 'Best moves:'
-        for b in n.best_moves:
-            print b
-            print
-        break
-
-
-if __name__ == '__main__':
-    main()
