+++ animal_shogi/exceptions.py
@@ -1,0 +1,6 @@
+class AnimalShogiException(RuntimeError):
+    pass
+
+
+class ParsingError(AnimalShogiException):
+    pass
